requirements.txt
amqp==5.2.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
black @ git+https://github.com/psf/black@ba88fc372eaed34abb33ea02d6860b185388d928
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
decorator==5.1.1
Django==4.2.11
django-extensions==3.2.3
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
environ==1.0
executing==2.0.1
idna==3.7
ipython==8.12.3
jedi==0.19.1
kombu==5.3.7
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.2.1
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.31.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
